+++ mafiaTelegramBot/src/main/java/co/mafiagame/bot/handler/StartCommandHandler.java
@@ -124,7 +124,7 @@ public class StartCommandHandler extends TelegramCommandHandler {
 
     private void readCitizenNo(StartGameState startGameState, TMessage message, MessageHolder.Lang lang) {
         try {
-            startGameState.citizenNo = Long.valueOf(message.getText());
+            startGameState.citizenNo = Integer.valueOf(message.getText());
             client.send(new SendMessageWithForceReply()
                 .setReplyMarkup(new TForceReply())
                 .setReplyToMessageId(message.getId())
@@ -141,7 +141,7 @@ public class StartCommandHandler extends TelegramCommandHandler {
 
     private void readMafiaNo(StartGameState startGameState, TMessage message, MessageHolder.Lang lang) {
         try {
-            startGameState.mafiaNo = Long.valueOf(message.getText());
+            startGameState.mafiaNo = Integer.valueOf(message.getText());
             client.send(new SendMessageWithReplyKeyboard()
                 .setReplyMarkup(new TReplyKeyboardMarkup()
                     .setKeyboard(Collections.singletonList(Arrays.asList(
@@ -202,8 +202,8 @@ public class StartCommandHandler extends TelegramCommandHandler {
     }
 
     private static class StartGameState {
-        private Long citizenNo;
-        private Long mafiaNo;
+        private Integer citizenNo;
+        private Integer mafiaNo;
         private Boolean hasDetective;
         private Boolean hasDoctor;
     }
